requirements.txt
selenium
undetected-chromedriver
mysql-connector-python
psycopg2-binary
pandas
pyspark
unidecode
apache-airflow
confluent-kafka
Flask-AppBuilder
SQLAlchemy
Flask
pyarrow
python-dateutil
Jinja2
marshmallow
numpy
Werkzeug
PyJWT
requests
pytz
urllib3
sqlparse
prison
PyYAML
SQLAlchemy-Utils
WTForms
Flask-Caching
click
setuptools
colorama
apispec
rich
MarkupSafe
croniter
pillow
cryptography
typing_extensions
simplejson
pyparsing
Flask-Login
Flask-JWT-Extended
babel
cron-descriptor
alembic
packaging
importlib-metadata
protobuf
python-dotenv